Resolução de exercícios da disciplina Estruturas de Dados 1

+++ src/ex56/Ex56.java
@@ -5,10 +5,11 @@ public class Ex56{
     //Algoritmo que pega nome e idade do usuário e os mostra a ele
     public static void main(String[] args){
 	String nome;//Variável usada
-	int idade;//Variável usada
+	byte idade;//Variável usada
 	nome = JOptionPane.showInputDialog("Digite nome");//Pegando o nome do usuário
-	idade = Integer.parseInt(JOptionPane.showInputDialog("Digite idade"));//Pegando sua idade e convertendo para tipo inteiro
+	idade = Byte.parseByte(JOptionPane.showInputDialog("Digite idade"));//Pegando sua idade e convertendo para tipo inteiro
 	JOptionPane.showMessageDialog(null,"Nome = "+nome);//Mostrando seu nome
-	JOptionPane.showMessageDialog(null,"Idade = "+idade);//Mostrando sua idade
+        JOptionPane.showMessageDialog(null,"Idade = "+idade);//Mostrando sua idade
+    	
     }
 }
